poc: Core controlled writes
In core controlled write, the feature serializes the message and send the result to the core to queue and cache / write to storage for eventual upload. Features have to keep track of current context and are informed of context changes through the same messages (as context changes tend to take place as part of something being written in RUM).