feat: Add storage and file wrappers (#7)
* feat: Initial pass at FileSystem wrapper

Datadog will need to access the filesystem to cache data as part of assembling batches to send to intake.  This is a first pass at filesystem access that is user overridable.

feat: Initial pass at FileSystem wrapper

Datadog will need to access the filesystem to cache data as part of assembling batches to send to intake.  This is a first pass at filesystem access that is user overridable.

---------

Co-authored-by: Steven Sudit (DataDog) <steven.sudit@datadoghq.com>