feat: Move storage writes to a background thread
The core now uses a background thread to write items to storage, using a Blocking Queue to support waiting for new messages. The new blocking queue only supports movable objects, so no copying of objects or additional allocation of memory is necessary as the objects move from one thread to another, and get written to storage.

refs: RUM-6940