chore: Refactor core to public and private interfaces
While implementing the reporting layer, I realized I was avoiding adding useful methods to Core that would allow us to split off some of the Core's internals into classes with single responsibilities, only because I didn't want to pollute the public interface with getters that were only for internal use.

To fix this, I've changed `DatadogCore` to be a (mostly) public interface for clients, created another interface for our internal methods (`DatadogCoreInternal`) and create the full implementation in a file under the internal namespace (`DatadogCoreImpl`).  This should allow us to create some accessors that we don't want being used by clients but can be used internally (such as accessing FeatureStorage).