feat: Add reporter and reporting thread
This is the skeleton of the reporting system for the core.

On wake up, the reporting thread processes a single "frame", which for each feature loads a configurable number of batch files, serializes them to a `Report`, then sends the `Report` to a `Reporter` (which will eventually be an upload to intake).

Unlike other SDKs, this core uses one thread for all reporting, instead of one thread per feature. It abandons reporting if a feature returns an error (assuming the error is in all of intake) deletes files directly from feature storage (instead of broadcasting a "batch closed" event).

Several features are still missing from the reporting thread core, including retry backoff, which I will make separate tickets for.

refs: RUM-6108