feat: Initial Libcurl reporter properly uploading to Datadog intake
This required sending `CoreContext` into the Feature in order to properly get all of the expected parameters needed for serializing the report.

There are a few places where we are generating a lot of temporary allocations to assemble strings. We will probably want to cache these in the long term, and look into ways to avoid passing unneeded headers from the feature to the reporter.

refs: RUM-6106