feat(phase-0): Update build configuration etc. for refactor
refs: RUM-11205

This is the first in a series of PRs to bring the changes I've been
working on for the past couple of weeks into the mainline branch of
`dd-sdk-cpp`.

The root ticket for this effort is
[RUM-11205](https://datadoghq.atlassian.net/browse/RUM-11205);
subsequent PRs will be tagged with additional Jira IDs as appropriate.

Once again, sorry for building up a massive changeset over two weeks
rather than merging to main incrementally - I'm sure I'll be much
better-equipped to tackle things piecemeal now that I have my bearings
in the project.

This PR is open for merge into a parent branch,
`aforsythe/RUM-11205/refactor`, which is in sync with `main`. Subsequent
PRs will be stacked incrementally on top of this one.

Therefore, review requests are not blocking: feel free to jump around or
let them sit for a while.

I'll take note of any significant issues raised in the review process,
and when it's time to merge the `refactor` branch into main, I'll
summarize those issues and note the action taken for each item (e.g.
fixed; not fixed but Jira ticket created; or no action taken).

Of course, if any issues demand immediate fixes, just flag your review
as "requesting changes" and I'll do my best to address them. And as
always, I welcome any and all questions and concerns: if I've thrown any
babies out with the bath water, let me hear it.

This PR mainly targets top-level files that deal with the development
environment and the project's tooling and build configuration. It does
not include source changes.

`clang-format` is still supported and mandated: auto-formatting über
alles.

I _have_ made some subjective changes to the configuration, so that the
style of the codebase will look a bit different. I _do_ believe that the
updated style is more readable and easier on the eyes, and it's also
closer to the style used in Unreal Engine (to ease the transition if
we're ever jumping between the two), but... of course it's largely a
matter of personal preference.

Reviewing more closely, I notice that we currently have `clang-format`
incorporated into the CMake build, whereas I'm running it in a simpler
script (e.g. in CI, `./build.sh --validate-format --rebuild` would
handle the whole build). If we want formatting to be part of the CMake
build, I can address that discrepancy.

Static checks with `clang-tidy` are still very much part of the build.
I've added a few sets of checks that weren't yet present, but I've also
suppressed a wider set of specific checks on a project-wide basis. I've
documented the reasons in the config file, and I'm happy to elaborate on
those reasons (or reconsider them) as needed.

All `clang-tidy` warnings are treated as errors. Checks run as part of
the CMake build.

Test-only binaries are not subject to `clang-tidy` checks.

I've temporarily removed the CI config, just so we don't waste CI
resources on this feature branch, which may not be in a buildable state
at certain points in its code-review journey.

I'll need to restore a working CI pipeline prior to the final release to
`main`.

I've removed the submodule config, as it seems easier to manage all
external dependencies via CMake, if we can get away with it.

As for the three specific dependencies that were managed as submodules:
I don't actually have them added back to the build via CMake yet, for a
couple of reasons:

- `date` and `rapidjson` are in the logging implementation (or the tests
  thereof), which is the last thing I have left to reconcile with my
  changes, so they'll probably be added back shortly.

- So far I've managed to author tests without needing `tromploeil`- but
  if we end up needing a mocking framework, I'll add it right back.

Sidenote: I've currently got tests (not part of this PR) at 86% function
coverage and 70% line coverage, and growing. That's involved plenty of
mocking, but I've found manually-written mocks to be very effective for
the sort of tests I've been writing. More on that when we get to code
changes.

Incidental changes. If any of this stuff should be left out of the
project-level config to be overridden on a per-user basis, we can figure
that out as we go.

Root build rules for the project. This PR does not include the
`CMakeLists.txt` files for specific targets; those will be up for review
along with their respective source.

Major build options at project-level:

- `DD_BUILD_EXAMPLES`: include example programs in the build

- `DD_BUILD_TESTING`: include unit test binaries in the build

- `DD_ENABLE_CLANG_TIDY`: If enabled, clang-tidy will be found and used
  for relevant targets. If clang-tidy is not found when this option is
  set, the build will fail.

- `DD_ENABLE_COVERAGE`: If enabled, test binaries will be instrumented,
   and a `coverage` target will be added that uses `llvm-cov` and
   `llvm-profdata` to analyze test coverage. If this option is set and
   either those LLVM tools are not found or the active compiler is not
   Clang, the build will fail.

Default options are set such that they'd make sense when the project is
included as a third-party dependency in someone else's project: i.e. we
don't enable tests/examples/checks/coverage by default.

As a convenience, setting `DD_DEVELOPMENT` will enable a reasonable set
of default options for library development in this repo: i.e. it'll set
`DD_BUILD_EXAMPLES`, `DD_BUILD_TESTING`, and `DD_ENABLE_CLANG_TIDY`.

This is a quick-and-dirty convenience script with two intended purposes.

One is to make it easier to run common build tasks without having to
`mkdir`/`cd`/`rm -rf` around and remember the right CMake incantations.
For example:

- `./build.sh` (from a clean repo) to run an incremental build
- `./build.sh --rebuild` to blow away all build state and start over
- `./build.sh --format` to `clang-format` the whole codebase and exit
- `./build.sh --test` to run unit tests
- `./build.sh --coverage` to run tests with coverage and open a report

The other purpose is to facilitate reproducible builds in CI, where we'd
want to strictly validate compiler versions and the like, e.g.

- `./build.sh --validate-format --compiler clang --clang-version 17`

This script makes a very basic attempt to facilitate builds with
different toolchains on different platforms, but that support is a
best-effort, untested first attempt: that's not the current focus.

`catch2` is still present as a test-only dependency.

`libcurl` is present as well, included when the default HTTP client
implementation is enabled (as it is by default; but that source and
config are not in this PR).

`expected-lite` is new: it's a header-only implementation of C++20's
`std::expected`, which is handy as a return type for functions that need
to signal failure. It's used within the implementation layer (also not
in this PR); it does not bubble up to the API layer.