feat(phase-1): Establish clearly-delineated C and C++ APIs
refs: RUM-11310

h## Rationale

The existing `dd-sdk-cpp` codebase deals largely with C++: it does not
yet have a C API.

Additionally, while parts of the existing `core` module are nested in an
`internal` namespace, with the `Core` itself using an inheritance
hierarchy to separate public vs. private details, the architectural
boundary between user-facing API surface and internal implementation
details could be much clearer.

RUM-11310 represents the effort to land on a design that:

- Makes it abundantly clear which parts of the codebase are private and
  which parts are public

- Prevents changes to private code from impacting ABI stability, as a
  default assumption

- Retains an idiomatic, user-facing C++ API that allows customers to
  integrate the SDK into their C++ projects, and that would also
  facilitate building other C++-focused products on top of this SDK

- Also exposes an idiomatic C API that would allow customers to
  integrate the SDK into their C projects, and that would also
  facilitate the creation of FFI bindings for other languages/runtimes

As discussed in https://datadoghq.atlassian.net/wiki/spaces/RUMP/pages/5383880755/Conventions+for+C+SDK+development,
I believe that a layered architecture with two independent APIs, one C
and one C++, is the sanest approach for us to take.

[image]

h## Changes in this PR

This PR includes:

- Example programs, in both C and C++, demonstrating usage of both APIs
- Public headers for the C API
- Public headers for the C++ API
- Source for the C API implementation
- Source for the C++ API implementation
- CMake configuration for examples and API bindings

This PR does _not_ include any of the implementation-layer source
referenced at the API binding layer. Those code changes will follow in
later PRs.

h### Caveats

One notable omission: the super-cool copy-on-write attribute system is
not yet represented in the logging API, as logging with arbitrary
attributes is the last bit of SDK functionality left to work on
in this refactor. (It now needs to be usable from both C and C++.)
That'll be addressed in a later PR.

Top-level SDK configuration is currently handled via structs that are
exposed directly to the user: `dd_core_config_t` in C and `CoreConfig`
in C++. Exposing the struct layouts publicly isn't ideal, as it makes it
onerous for us to change these types, and it also doesn't let us
validate user-supplied arguments. In future work, I intend to replace
the C struct with an opaque type and option-setter functions (e.g.
`dd_config_set_client_token(dd_config_t* config, const char* value)`),
and I intend to replace the C++ struct with a `ConfigBuilder` type.