feat(phase-3): Make HTTP client implementation portable
refs: RUM-11312, RUM-11315, RUM-7617, RUM-7075, RUM-11205

This PR gives our HTTP client implementation a similar treatment as
filesystem access, making it a portable "subsystem" within the platform
abstraction layer, facilitating alternative implementations at
build-time, and improving test coverage.

- RUM-11312 represents that refactoring effort
- RUM-11315 for test coverage
- Also addresses these existing issues:
    - RUM-7617: Set up libcurl init and cleanup to be static
    - RUM-7075: Provide a way to remove libcurl

Behavior of the libcurl implementation is largely unchanged, with one
notable exception: we now use `CURLOPT_READFUNCTION` to facilitate
zero-copy transfers.

Previously, uploading a batch of event data required building the full
HTTP request body in-memory, meaning that the worst-case memory overhead
for that operation was effectively the maximum batch size (4 MB).

Using `CURLOPT_READFUNCTION`, we can stream event data directly to the
HTTP connection as we read it from a batch file. For event formats like
logging, where we don't do any filtering and we simply concatenate the
events from the file into a JSON array, this effectively reduces
worst-case memory overhead to the maximum size of an _event_. (Features
that filter and/or transform event data based on metadata may be less
able to take advantage of this optimization, and they can opt out of it
by precomputing the full payload prior to initiating the request.)

With HTTP/1.1, libcurl uses `Transfer-Encoding: chunked` for these
requests.

h## Code changes

In `src/impl/`:

- `platform/http.hpp` defines platform-agnostic interfaces
    - `IHttpClient` is an HTTP client: given the usage patterns of the
      SDK, we assume all requests are POSTs that return no meaningful
      response body
    - `IHttpSubsystem` provides a place for global init/shutdown logic,
      in the event that we create multiple clients (currently the SDK
      just uses a single HTTP client)
    - Error semantics are slightly different: the HTTP client
      implementation is not responsible for classifying HTTP responses
      as retryable/non-retryable based on status code (that still
      happens, but at the application layer); it just concerns
      itself with errors that are internal to the client
    - `HttpBodyWriter` denotes a callback that populates an HTTP request
      body; subsequent changes include ready-made implementations
      (`StringWriter` and `TLVBatchWriter`, not in this PR) so that
      features don't have to deal with HTTP-request-body-chunking on
      their own
- `platform/http_curl.cpp` is the libcurl implementation
    - Note that this implementation would likely not be compatible with
      running multiple Datadog SDK instances in the same process: we'd
      want to add a reference count guarding `curl_global_init` and
      `curl_global_cleanup` if that's a use case we want to support
- `CMakeLists.txt` includes `http_curl.cpp` in the build by default,
   also including libcurl as a dependency iff using the default
   implementation; you can use a different implementation by disabling
   `DD_HTTP_USE_DEFAULT` and supplying `DD_HTTP_SRC`

In `tests/`, to validate our libcurl implementation (or any other
implementation that we or a user might write):

- `support/http_server.hpp` implements a bare-bones, in-memory HTTP
  server that wraps a TCP socket, records incoming requests, and sends
  dummy responses, to help test HTTP client functionality with a real
  connection
- `impl/platform/http_test.cpp` verifies that the HTTP client
  implementation works as intended

Finally, we have another set of mocks for use in tests elsewhere in the
codebase:

- `mock/http_client.hpp` implements `MockHttpClient`, which caches the
  full details of every outgoing request, along with the details of any
  accompanying response
- `mock/http_client_mock_test.cpp` verifies that the mock client works
  as intended