feat(phase-4): Make clock a platform subsystem
refs: RUM-11205

This PR is a bit simpler than the last two: it adds our third and final
(for now) platform subsystem: the clock.

This interface, `IClock`, simply abstracts access to the system clock.
This doesn't technically need to be a class, since it's only one
dead-simple operation (reading the clock) that doesn't require any
accompanying data. However, I chose to turn it into a full-fledged
interface for consistency with the rest of the platform layer: it's
constructed and injected into the core in the same way, with the same
ownership semantics.

Having a tiny bit of indirection on clock reads (rather than simply
calling `std::chrono::system_clock::now()` directly) is handy for
testing, but this also allows us to plug in a different clock
implementation at build time.

We're targeting C++17, so we _can_ assume `std::chrono` will be
implemented on all target hardware, but the standard doesn't make any
guarantees about the precision of `system_clock`, so presumably we're
not taking it on faith that all platforms' implementations of
`std::chrono` will suffice for our needs.

Note that we previously cast values read from the clock to a raw
`uint64_t`: it seems preferable to use `std::chrono` types so as to
retain the type-safety they provide. I've chosen to store durations and
timestamps from the platform clock interface as _signed_ `int64_t`, as
it allows negative deltas, and the year 2262 is pretty far away.

Code changes:

- `clock.hpp` is the interface
- `clock_std.cpp` is the std::chrono implementation
- `clock_test.cpp` sanity-checks basic invariants
- `mock/clock.hpp` is a mock implementation for use in tests