feat(phase-6): Refactor storage thread, handle tracking consent
refs: RUM-11313, RUM-11316

h## Introduction

This PR refactors the implementation of the storage thread so that:

1.) It uses the `platform/filesystem` interface
2.) It accounts for the need to handle event data differently based on
    tracking consent
3.) It supports event metadata

Not done in this PR:

- Deletion and migration of pending event data on consent change
    - The SDK writes event data to the appropriate subdirectory (or not
      at all) depending on current tracking consent, and if tracking
      consent changes at runtime it will update its behavior, but it
      does not yet delete or move the files in the pending-events subdir
- Eager deletion of batches older than the max allowed age (18h default,
  configurable per feature)
    - The mobile SDKs appear to do this in both the storage and upload
      threads, which would require
- Batch metadata
    - We now support writing per-event TLV metadata blocks; we do not
      yet support storing whole-batch metadata it separate files

h## Changes in this PR

The code used by the storage thread is defined in `storage.hpp`:

- `StorageMessage`, a type representing messages that can be sent to the
  storage thread via a queue
- `BatchWriter`, an interface that commits event data to disk, keeping
  track of batch files etc.
- `EventStorage`, which sits in front of two `BatchWriter`s and routes
  events appropriately based on tracking consent
- `StorageThreadMain`, the entry point for the storage thread, which
  operates on some state that's passed in by the Core

This code is an iteration on `feature_storage.h` et al., with a few
tweaks:

- Writing and reading are separate concerns; the code that writes batch
  files is separate from the code that reads batch files
- The storage thread does not directly access the core
- The standard code path for writing events is lock-free

h## Context re: the Core's threading model

The `Core` implementation is not part of this PR, so here's a bit of
context about the core as it relates to the storage thread.

The Core owns a `std::vector<RegisteredFeature>`, with one item for each
registered feature. The Core also owns the storage thread, along with
the thread-safe queue used to pass messages to that thread:

```cpp
class Core
{
    // ...

    std::vector<RegisteredFeature> _features; // Immutable while running

    // Initialized on Start, cleaned up on Stop
    std::unique_ptr<Queue<StorageMessage>> _storage_queue;
    std::optional<std::thread> _storage_thread;

    // ...
};
```

`RegisteredFeature` records each feature's ID and name, and it also owns
a bunch of feature-specific state: some exclusive to the main thread,
some exclusive to the storage thread, and some exclusive to the upload
thread. As far as the storage thread is concerned, these are the only
members that matter:

```cpp
/** State maintained by the Core for a feature that's been registered with it. */
struct RegisteredFeature
{
    /** FourCC ID that uniquely identifies this feature. */
    FeatureId id;  // `using FeatureId = uint32_t;`

    /** Unique name of this feature, used in logs, directory names, etc. */
    std::string name;

    // ...

    /** Used by the storage thread to commit event data to disk. */
    std::unique_ptr<EventStorage> event_storage;

    // ...
};
```

This is the most here-be-dragons part of the implementation layer:
`RegisteredFeature` is where we centralize the state that's accessed by
different threads, and we use a lock-free design, so we need to be very
deliberate about our thread-safety invariants.

On `Core::Start()`, the core creates the storage thread via
`std::thread`. On `Core::Stop()`, the core joins on that thread. Each of
these operations is a synchronization point.

As of the thread-start synchronization point, the core effectively
grants exclusive access to all features' `EventStorage` structs to the
storage thread: it guarantees that no other thread will write to or read
from `EventStorage` for as long as the thread is alive.

Furthermore, the Core guarantees that the  `_features` vector will
remain entirely immutable for as long as the core is running (and
therefore, for the lifetime of the storage thread): nothing will mutate
either the vector or its items.

This makes it safe for the storage thread to accept a
`std::vector<RegisteredFeature>&` parameter, allowing it to iterate over
the core's features and dispatch function calls to the relevant
feature's `EventStorage`.

Basically, this is what the storage-related state looks like once a
feature is registered:

[image]

The Core creates the RegisteredFeature (and all its members) when the
feature it registered; and then it never touches it again.

Then, on core start, we do this:

```cpp
_storage_queue = std::make_unique<Queue<StorageMessage>>();
_storage_thread = std::thread(
    StorageThreadMain,
    std::ref(*_storage_queue),
    std::ref(_features)
);
```

...so we end up with:

[image]

For as long as the core is running, the only way that data is sent to
storage is via the thread-safe storage queue.

h## Summary

The entry point of the storage thread is `StorageThreadMain`. All that
function does is continually read from a blocking, thread-safe queue
which can contain two types of messages: `TrackingConsentChanged` or
`EventGenerated`.

On `TrackingConsentChanged`, the storage thread will call
`EventStorage::SetTrackingConsent` for all features. On
`EventGenerated`, the storage thread will call
`EventStorage::HandleWrite` for the feature that generated the event.

`EventStorage` sits in front of two `BatchWriter` instances, one for the
directory containing tracking-consent-pending events, and the other for
the directory containing tracking-consent-granted events. `BatchWriter`
handles the logic related to batching events into files and writing
changes to disk; `EventStorage` delegates between the two based on
tracking consent.

When the core is stopped (i.e. on shutdown), queue processing stops, and
the storage thread is allowed to continue running until it drains the
queue. Once that's happened, the storage thread will exit.