feat(phase-9): Refactor Attribute implementation
refs: RUM-11313, RUM-11316

This PR iterates on the original `DatadogAttribute` implementation,
replacing it with a new version that is usable from both C and C++ APIs.

h## Changes to production code

Here's a rough description of how the runtime code related to attributes
is now organized:

[image]

h### The C API

The C API defines a struct `dd_attribute_t` which is returned by value,
along with a series of C-style API functions used to operate on those
values, with the following mandates:

- New `dd_attribute_t` values must be obtained by calling
  `dd_attribute_<type>()`; existing values may be reinitialized by
  passing them to `dd_attribute_set_<type>()`.
- `dd_attribute_t` must be copied via `dd_attribute_copy()`.
- All `dd_attribute_t` values must be passed to `dd_attribute_free()`
  before leaving scope.

The final test in `attribute_c_api_test.cpp` shows how this API can be
used to construct an arbitrary object value.

h### The C++ API

The C++ API defines `datadog::Attribute`, which is also returned by
value. Thanks to RAII, the C++ interface is much more straightforward to
use. See `attribute_cpp_api_test.cpp` for an equivalent object-building
example.

h### Separation of Attribute (API) and CowValue (impl)

Compared to the core initialization and feature registration APIs, the
attribute API necessarily exposes a lot more details publicly: this
allows the C and C++ attribute APIs to use idiomatic constructs from
their respective languages, while ensuring that attribute values can be
allocated and stored in application-space. As such, there's a fair bit
of duplicated logic between the C and C++ APIs.

However, the "Attribute" type (in either API) is now solely responsible
for providing the user-facing interface, storing trivially-copyable
primitive values, and wrapping the underlying `CowValue` pointer used
for non-primitive values. All of the reference-counting and
copy-on-write behavior is now confined to the implementation layer, in
the `CowValue` type.

Within the implementation layer, we use the C++ `Attribute` class
exclusively: for example, a `CowValue` that holds array items will store
those values in a vector of `Attribute`.

`AttributeConversion` defines helper functions to perform lightweight
copies between `Attribute` and `dd_attribute_t` (with proper reference
counting) across the C API boundary.

h### Notable changes to CowValue

I've attempted to nail down a few details of our copy-on-write
implementation while respecting the design of the original
`DatadogAttribute`. These are the key things I saw fit to change, in
addition to the layered design with cleaner separation of concerns
between attributes and values:

- Reference counting is now implemented directly, with a
  `std::atomic<int>` in each `CowValue`, rather than implicitly relying
  on the behavior of `std::shared_ptr`

- An `Attribute` stores a type tag and a union of _either_ a 64-bit
  primitive value _or_ a raw `CowValue*`, rather than storing a
  `std::shared_ptr<CowStorage>` _alongside_ the primitive value

    - `sizeof(Attribute) == 16`

- A `CowValue` stores a type tag and a union of either `std::string`,
  `std::vector<Attribute>`, or
  `std::vector<std::pair<std::string, Attribute>>`, rather than relying
  on raw pointers and manual memory management via `new[]` and
  `delete[]`

    - `sizeof(CowValue) == 32`

    - String values may allocate additional heap memory to store their
      value, but relying on `std::string` means that we now get small
      string optimization (SSO) for free if enabled: values under the
      SSO threshold (typically ~22 chars) fit within the `std::string`
      value itself.

    - Array values allocate an additional 16 bytes per item.

    - Object values allocate an additional 40 bytes per property. If SSO
      is in use, short property names will require no additional heap
      allocations.

h### JSON Serialization

Finally: our existing implementation had basic support for serializing
Attribute values to JSON. It's actually not that crazy to roll our own
implementation here, given that a.) any Attribute value is effectively a
well-formed AST of a JSON value, and b.) we only need to serialize _to_
JSON, which is fairly straightforward (compared to the nightmarish
complexity of _parsing_ JSON).

Serialization of Attribute values is now implemented in an
`AttributeSerialization` helper, defined at the implementation layer and
declared as a `friend` of `Attribute` (as JSON serialization is a
tangential bit of functionality that's only used at the implementation
layer; it's not provided as part of the public API).

I've tried to make this implementation fully robust and
production-ready:

- It now properly escapes string values, to ensure that strings with
  quotes, backslashes, and ASCII control codes will be handled properly

- It now coerces non-finite double values to `null`, rather than
  emitting `nan`, `inf` or `-inf`, which are not valid JSON literals

- We now traverse the object once to precompute the required buffer size
  before serializing

h## Tests

[image]

Here's a rundown of the relevant tests that I've added.

- `c/attribute_c_api_test.cpp`: fully exercises the C API, including
  edge cases that are specific to C (e.g. null pointer/string checks),
  and validating end-to-end that an object built in C is serialized to
  JSON as intended

- `c/attribute_cpp_api_test.cpp`: fully exercises the C++ API, with the
  same end-to-end serialization test

- `impl/attribute/cow.cpp`: unit-tests the `CowValue` implementation in
  isolation

- `impl/attribute/json_test.cpp`: unit-tests the
  `AttributeSerialization` implementation, to ensure that every
  imaginable JSON edge case is handled

- `impl/attribute/memory_test.cpp`: profiles memory usage to codify the
  expected behavior re: when and how Attributes allocate heap memory

- `impl/attribute/refcount_test.cpp`: examines the internal state of
  Attribute values to verify that reference counting and copy-on-write
  work as intended

These latter two tests rely on a few support files:

- `support/allocation_tracker.hpp`: defines overloads of `new` and
  `delete` that instrument allocations/frees for the scope of a test,
  then provide summarized stats for examination at the end of the test.
  Can be disabled (to ensure that tests use unmodified `new`/`delete`)
  by disabling `DD_ENABLE_TEST_ALLOCATION_TRACKING`.

- `support/attribute_debug.hpp`: defines another friend helper struct,
  `AttributeDebug`, which contains utility functions used to examine the
  internal state of Attribute values in test-only code.

- `support/memory_checks.hpp`: examines the build configuration to
  determine whether `WITH_DATADOG_STRICT_MEMORY_CHECKS` should be
  enabled: this allows us to write tests that rely on compiler-specific
  behavior (e.g. SSO, STL allocation strategies), without getting false
  positives in builds on different platforms.

I'll admit that I was having a little bit of fun with these memory
tests: it may make sense to pare them back a bit once we have proper
profiling and benchmarking tools in our CI workflow.