chore: Reconcile stray source files with no-op changes
I was originally working in a separate repo and pulling in changes
to `dd-sdk-cpp` piecemeal, and after all my manual copying, this small
set of files was slightly out of sync. This PR just merges the
inconsequential changes that I missed in earlier PRs, so that I can drop
the WIP repo for good.

These are all whitespace-only changes that resulted from
auto-formatting, plus one fix to a misnamed `#include`. (Actual
re-formatting pass TBD.)