feat: Allow pending batches to be flushed via HTTP on shutdown
refs: RUM-11205

This PR adds a config option called
`num_http_requests_per_feature_to_flush_on_stop`: if nonzero, when the
Core is stopped, it will synchronously initiate an upload cycle for each
registered feature, uploading up to N batches for each feature. These
final uploads occur in the main thread, after the Core has joined on
both the storage and upload threads, ensuring that state is
synchronized and that all batch files are safe to read.

h### Caveat: Flush-on-shutdown is for tests; public API (if any) TBD

This change is primarily intended to facilitate unit testing. Whether we
want to expose this feature to users (and if so, what level of control
we'd want to offer them) is an open question.

The design of the `CoreConfig` and `dd_core_config_t` structs is not
final. In a future change we'll almost certainly update the core
configuration API so that it's less brittle w.r.t. newly-added fields
(using a builder type in C++, and using property-setter functions in C),
and in that process we can decide on one of these two courses of action:

1. Make flush-on-shutdown an officially-supported, user-facing feature,
   determine the extent to which it should be user-configurable, and
   ensure that the config API expresses the option in those user-facing
   terms

2. Make flush-on-shutdown an internal, test-only feature, and ensure
   that it's configurable in tests but not exposed directly to users via
   the public config API

h### Why this feature is useful for tests

We want to be able to write tests that exercise some feature API or
another, then validate that the expected HTTP requests were sent. By
processing all pending batches for upload on Core shutdown, we ensure
that our tests can immediately inspect the set of events that were
generated, including the way in which they were processed and batched
for upload.

h### New test coverage

New tests in `feature_test.cpp` demonstrate this pattern. By default,
`CoreTestHarness` now configures the core to flush _all_ pending
requests on shutdown. So we can now validate the event-handling pipeline
all the way through, from generation of individual events to processing
and upload of the resulting batches.

h### Thread-safety tests

This PR also adds an initial test with the `[thread-safety]` tag. These
tests aim to validate thread-safety guarantees.

We can probably have a larger discussion about thread safety, but here's
my perspective as it relates to testing: regardless of what kind of
thread safety guarantees we do or don't offer in the C++ SDK, we should
be explicit about those guarantees. This requires us to write tests that
are intended to provoke race conditions, so we can guarantee that none
occur in our supported use cases.

This mostly comes into play when we talk about individual feature APIs,
but I've added a test to the core event-processing loop in this PR. I've
also added a `RunParallel` function in `support/threading.hpp` to try
and make these thread-safety tests easier to write.