feat: Refactor logging implementation, add sampling/filtering
refs: RUM-11314

This PR updates the existing logging feature implementation to support
both C and C++ APIs, while also adding support for the "remote log
threshold" and "remote sampling rate" parameters. It also expands test
coverage, ensuring that both APIs are exhaustively tested and
validating our (rather minimal) thread-safety guarantees.

To be explicit about those thread-safety guarantees: as currently
implemented, it _is_ safe to create and use multiple loggers
concurrently from different threads. It is _not_ safe to use the _same_
logger concurrently from different threads.

h## Code changes

h### Logging API

`datadog/logging.h` and `datadog/logging.hpp` define C and C++ APIs,
respectively. These headers were included in a previous change; as of
this PR they are fleshed out and fully documented.

In this PR, the key change to these APIs relates to logger
configuration. Previously, both APIs exposed structs for the user to
populate directly, a pattern which hampers our ability to validate
arguments and which creates confusion about which fields are required
etc. (in addition to ABI compatibility concerns)

As of this PR, the C++ API still exposes a struct type, but its fields
are now private. A default-initialized `datadog::LoggerConfig` will have
sensible default values; individual setter functions can then be called
(or chained) to customize the configuration.

The C API now uses an opaque struct: under the hood,
`dd_logger_config_t` simply holds a `datadog::LoggerConfig` by value.
Similarly, `dd_logger_config_create()` creates a default-initialized
config, and setter functions (e.g. `dd_logger_config_set_service()`)
simply call the underlying setters on the C++ type.

This approach _does_ require the C API (though not the C++ API) to
dynamically allocate (and later free) its custom logger config structs.
We could avoid that allocation, but it would require duplicating the
config type wholesale (so its members could be exposed in the C API)
rather than simply wrapping the C++ type under the hood.

As with the core, the `logging` and `logger` calls for both APIs defer
to an underlying `impl::Logging` and `impl::Logger`, with the C and C++
APIs being more-or-less independent.

h### Minor core changes to support logging feature

A feature implementation needs a limited set of information from the
Core. In the case of the logging feature, we need access to the system
clock so that we can timestamp our log events, and we need basic global
state like the configured service name and application version.

Here's how I've chosen to handle this for now: rather than having a hard
dependency between the feature implementation and the core, each
`impl::Feature` defines whatever state it needs via its constructor. The
API binding layer (i.e. `src/c/` or `src/cpp/`) is responsible for
injecting those dependencies when initializing the feature, prior to
registering it with the core.

To that end, I've added const accessor functions to `impl::Core` (not
the API-layer `Core`) that expose the state required by the logging
implementation: `GetClock()`, `GetServiceName()`, and
`GetApplicationVersion()`. These are injected on init, minimizing the
shared surface area between the core and the feature implementation.

h### Logging feature implementation

At the implementation layer, we have a `Logging` type (the feature
implementation), as well as `Logger` (which manages the state for a
single logger).

The logic related to generating log events needs to be centralized to
some degree, since we need to merge in global state from the `Logging`
feature. Therefore, when a `Logger` needs to emit a log event, it
ultimately defers to `Logging::OnLoggerEmit`. This function needs to
take a bunch of state, including some state that's owned by the logger
itself: namely, the reusable `Attribute` where we construct the event
payload, and the reusable buffer where that payload is serialized before
being copied into the storage queue. By maintaining this state
separately for each logger, we ensure that multiple loggers can emit log
messages concurrently.

When `Logging` constructs a `Logger`, it initializes the logger with a
callback that it can use to emit a log message (via `OnLoggerEmit`).
This callback captures a reference to the `Logging` instance via
`weak_ptr`, such that:

1. It's _always_ safe for the logger to invoke the callback, and
2. If a logger is used after the feature itself has been destroyed, it's
   simply a no-op

I'm open to reevaluating the exact mechanics of this dependency between
`Logger` and `Logging`. I _do_ like the callback-with-captured-weak-ptr
approach, as opposed to having a `Logger` directly access the `Logging`
instance, because I think it's best to avoid bidirectional coupling
wherever possible. However, I _do_ also find the sheer number of
parameters in the callback type a bit distasteful.

h### Test coverage

Tests are concentrated at the API layer, with some duplication:
`logging_c_api_test.cpp` and `logging_cpp_api_test.cpp` are essentially
identical; they just exercise different APIs.

It might be argued that some of this functionality could be unit-tested
in isolation at the implementation layer, but that would require a much
more complicated test design in order to cover both APIs as well, and
that additional complexity didn't seem warranted.

As it stands, these tests give us full confidence in the behavior of the
logging implementation when used with _both_ APIs.